287. Find the Duplicate Number
